+++ plugins/node/instrumentation-mongoose/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.42.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-mongoose-v0.41.0...instrumentation-mongoose-v0.42.0) (2024-09-02)
+
+
+### Features
+
+* update deps matching "@opentelemetry/" ([9fa058e](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/9fa058ebb919de4e2a4e1af95b3c792c6ea962ac))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.40.0 to ^0.41.0
+
+## [0.41.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-mongoose-v0.40.0...instrumentation-mongoose-v0.41.0) (2024-08-27)
+
+
+### Features
+
+* **instrumentation-mongoose:** Support v7 and v8 ([#2353](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2353)) ([770130a](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/770130a7a41125dbf8a8c18b3384e5fcc370deef))
+
 ## [0.40.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-mongoose-v0.39.0...instrumentation-mongoose-v0.40.0) (2024-07-03)